GOCBC-976: Allow WithExpiry to work with custom transcoders
Motivation
----------
When a user performs a Get WithExpiry we check to make sure that
no transcoder has been set. We shouldn't do this as this is a
valid usecase, we should allow transcoders to be used with this.

Changes
-------
If WithExpiry is used then:
Do not reject transcoders on the Options.
Fetch the flags via subdoc so that the transcoder has the correct
flags.
Update projections to allow transcoders which support JSON.

Change-Id: I830b5fec8ad16ede51ddccac07c0b0073e6ca328
Reviewed-on: http://review.couchbase.org/c/gocb/+/134737
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>