GOCBC-980: Update get projections with expiry result parsing
Motivation
----------
We introduced a bug when adding support for transcoders to get with
expiry. We only remove the expiry xattr op from the results content
if it was performed as a full get operation.

Changes
-------
Always remove the expiry xattr op from results content when doing
any type of get with expiry.

Change-Id: I948f2dc7fa8ce240c5fa0ac48d858ca09e6b44c6
Reviewed-on: http://review.couchbase.org/c/gocb/+/134972
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>