GOCBC-977: Alter signature of GetPendingMutations
Motivation
----------
How we parse the server response for GetPendingMutations is incorrect
and causes the unmarshaller to error if the response body contains
any data. This means that the signature of the function is also
incorrect. As no users can be using this functionality we should make
a breaking change to update the signature.

Changes
-------
Alter GetPendingMutations signature and update how we parse the server
response body.

Change-Id: Ic3c93a323cb75028963204fb40df6709837d43ce
Reviewed-on: http://review.couchbase.org/c/gocb/+/136388
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>