GOCBC-990: Update enhanced durability adaptive algorithm
Motivation
-----------
When the user uses enhanced durability we should send a value of
timeout*0.9 to the server. We actually send timeout*(10/9). When
durability is in use we should also coerce all timeouts up to 1500ms.

Changes
-------
Update the algorithm to apply timeout*0.9.
If enhanced durability in use and user set timeout is < 1500ms then
log a warning and coerce it to 1500.

Change-Id: I7b57761168aa970c9184beed51ae41db143ebeba
Reviewed-on: http://review.couchbase.org/c/gocb/+/136687
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>