GOCBC-994:  Convert subdoc nil values to json null
Motivation
----------
When using subdoc mutations if the user sets a value to nil then
the operation fails because we send no value to the server. Instead
of doing this we should coerce a nil value to a json null value.

Changes
-------
If a subdoc mutation value is set to nil then coerce it to a
json null.

Change-Id: I2ba84a91ff04973750aa322bf9c817553f210ffc
Reviewed-on: http://review.couchbase.org/c/gocb/+/137315
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>