GOCBC-996: Remove `[]` from id for `CollectionMap.At` and `Exist`
Motivation
-----------
The Map data structure is using invalid paths for At and Exists.
This is causing these functions to be unuseable. They wrap
the id in `[<id>]` which looks like a copy and paste bug from
the list/set data structures.

Changes
-------
Don't wrap the id in `[<id>]` for At and Exists for the map
data structure.

Change-Id: I5f7dfcb506c260603405e5b6b8e0eabcb703425b
Reviewed-on: http://review.couchbase.org/c/gocb/+/137041
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>