Update tests to not expect retries from Unlock
Motivation
----------
We have updated the gocbcore retry logic to match the spec so that errors that
we have knowledge of are not always exposed to the retry handler.
We only send a subset of errors to the retry handler and do not
retry, known, other errors. Gocb tests test that Unlock operations
are retried which is not correct behaviour.

Changes
-------
Remove timeout checks from Unlock test.
Change the timeout error test to use Upsert rather than Unlock.
Change-Id: I60ed723f7eb9e9522d439a5b4bee3ab53ffc2218
Reviewed-on: http://review.couchbase.org/c/gocb/+/140447
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>