GOCBC-948: Refactor document expiry duration
Motivation
----------
Several operations allow users to set an expiry duration for a
document, for durations of under 30 days this works as expected.
For durations of over 30 days the server interprets the value
as an absolute timestamp so will often immediately expire the
document if the expiry was set as a duration and not a timestamp.
We should add logic to detect if an expiry is >= 30 days and if
so then we should convert it to a timestamp using now + expiry.

Changes
-------
Add logic to detect if an expiry value is >= 30 days, if so then
apply the expiry as now in seconds + expiry value.

Change-Id: Ie9e6b84750ca5f5c98885dbf42bfb358672c01cf
Reviewed-on: http://review.couchbase.org/c/gocb/+/141335
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>