GOCBC-1019: Convert subdoc cas mismatch error to exists
Motivation
----------
In order to be consistent gocbcore now converts subdoc key exists
errors to cas mismatch, like it does for full doc. We cannot do this
as it would cause a backward incompatible change so we need to
convert the error back.

Changes
-------
Detect cas mismatch in mutate in and swap it for a document exists
error.

Change-Id: I2713988fbe11c3f90778893dd18b95f337e797a2
Reviewed-on: http://review.couchbase.org/c/gocb/+/141366
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>