GOCBC-1022: Route view queries to agents not agent group
Motivation
----------
At present we route view queries to the agent group ViewQuery
function which means that the queries are sent to whatever bucket
(if any) the last new config belonged to. This is an issue when a
user has multiple buckets connected and issues a view query as it
can go to the wrong bucket.

Changes
-------
Update connection manager so that get view query provider takes
a bucket name and returns the agent corresponding to that bucket
rather than the agent group.

Change-Id: Id0f11fc7c590b7cfa121d64017c0c2665e4079a9
Reviewed-on: http://review.couchbase.org/c/gocb/+/141443
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>