GOCBC-1021: Route bucket level http requests to agents not agent group
Motivation
----------
At present we route requests like view indexes managements to the
agent group HTTP handler function which means that the requests
are sent to whatever bucket (if any) the last new config belonged to.
This is an issue when a user has multiple buckets connected and
issues a bucket level http request as it can go to the wrong bucket.

Changes
-------
Update connection manager so that get http provider takes
a bucket name and returns the agent corresponding to that bucket
rather than the agent group if the name is not empty.

Change-Id: Ieb9438c816739be9faf36031201af72def7e0bd6
Reviewed-on: http://review.couchbase.org/c/gocb/+/141449
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>