GOCBC-1023: Add interfaces for accessing raw bytes from http services
Motivation
----------
Sync Gateway has expressed a desire to be able to access view rows
as raw bytes. This is so that they can unmarshal the entire row in
one go rather than needing to know the type of the key and the value.
We should allow this and also roll the same interface out to other
services.

Changes
-------
Add a volatile stability Raw API to query, search, views, and analytics.

Change-Id: I18d5f075b31c7b5998c5ecfcbc43757e9e7088ca
Reviewed-on: http://review.couchbase.org/c/gocb/+/141720
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>