GOCBC-1040: Include error causes when doing serialization
Motivation
----------
When we serialize our error types we omit the underlying "inner"
error from the serialized data. We should include this information
in the serialized data.

Changes
-------
Add custom MarshalJSON functions to our error types and convert
the errors to anonymous structs so that we can stringify the
"inner" error before marshalling the error.

Change-Id: I80ccf1dbb2b3abb902d5a1e040211b8122f98b9e
Reviewed-on: http://review.couchbase.org/c/gocb/+/144857
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>