GOCBC-1055: Add doc flags as an internal mutate in option
Motivation
----------
We have a set of subdoc flag constants but no way to use them.
Sync Gateway require a way to use CreateAsDeleted with mutate in.
Rather than add this as an explicit option we should remove
AccessDelete option and add a DocFlags option. This means that
they can use any flags that they need without requiring code
changes in the SDK, a kind of escape hatch.

Changes
-------
Remove AccessDeleted from the subdoc internal options.
Add DocFlags to the subdoc internal options.

Change-Id: I1f560dd51d28f3792fafc28e494f3ddafe092f76
Reviewed-on: http://review.couchbase.org/c/gocb/+/145440
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>