GOCBC-1063: Add support for context.Context
Motivation
----------
Usage of context.Context is standard within Golang for deadlines
and cancellation of requests. We previously chose not to support
this due to the level of complexity that it would add. We are
now in a position where we can support this without adding
a large amount of complexity and where we can add it without
modifications to gocbcore.
Support for context should break the standard idiom of it
being the first parameter in a function as this would require
releasing a new major version of the SDK or ballooning the API.

Changes
-------
Add context.Context as an option field in all option blocks for
functions which perform IO.
Update the "providers" which marshal between gocb and gocbcore
to support context.
Update the kv and async op managers to handle context Done signals
and cancel the underlying gocbcore requests.

Change-Id: I0d486f8fcee691c76316ab377db92c311fca6b38
Reviewed-on: http://review.couchbase.org/c/gocb/+/152907
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>