GOCBC-1044: Add metrics interface and support
Motivation
----------
As a part of updating observatibility support we are adding support
for metrics. We should provide a metrics interface and hook into
gocbcore's metrics interface, providing a wrapper. We should also
provide a requests metric for each call, which will record the
overall latency of an operation. We should also provide two metrics
implementations: noop and aggregating meters.

Changes
-------
Add metrics interface with default implementation of an aggregrating
meter, and an implementation of a noop meter.
Decorate every function call with a requests latency value record.

Change-Id: Ic1128798e1f7393a3bf83679bd65a8763fe85d83
Reviewed-on: http://review.couchbase.org/c/gocb/+/150779
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>