GOCBC-1070: Return bucket not found from flush when bucket does not exist
Motivation
----------
From bucket manager operations if the bucket specified does not
exist then we return ErrBucketNotFound, except for flush where we
return an unspecified error.

Changes
--------
Add a check to the flush error message parsing to detect if the
error was a 404 with resource not found, if so then return an
ErrBucketNotFound.

Change-Id: Ie24523c04f74709386e195813c28bbea7a632f6c
Reviewed-on: http://review.couchbase.org/c/gocb/+/148070
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>