Update core tracer wrapper to match gocbcore interface
Motivation
----------
As a part of tracing updates we have updated the gocbcore interface.
There are gocb side interface changes to come but the integration
of opentelemetry will require us to upgrade to a minimum supported
go version of 1.14 which will need a minor version bump. We aren't
quite ready to do that yet so we need to stopgap by updating our
gocbcore tracer wrapper.

Changes
-------
Update the core tracer wrapper to satisfy the new gocbcore tracing
interfaces.

Change-Id: I5ff1b381487ec9fc5406db3a4840e0801318ee29
Reviewed-on: http://review.couchbase.org/c/gocb/+/148191
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>