GOCBC-1074: Update threshold logger tag names
Motivation
-----------
Gocbcore recently updated all of the tracing span names and
tags to be up to date with the latest observability RFC. For
various reasons gocb has not yet been updated, we need to adjust
the threshold logger to look for the new span tags from gocbcore.

Changes
-------
Update the span tag names in threshold logger.

Change-Id: Ic25319da6a3037da49d4b9824dcd77e370b4ebe1
Reviewed-on: http://review.couchbase.org/c/gocb/+/149504
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>