GOCBC-936: Update analytics management API to support compound dataverse names
Motivation
----------
Dataverse names in 7.0 can now contain "compound" names. That is
names with parts which are separated by "/"s. We need to
detect this in analytics management functions which send analytics
queries and we should split on /s and replace them with .s, escaping
each part of the name.

Changes
------
Update the analytics management functions which are backed by analytics
queries to detect compound dataverse names and split and rejoin the
names with .s, escaping the parts.

Change-Id: Ia84709c02483b2c2f39cfff3c3cf78839fdbcb71
Reviewed-on: http://review.couchbase.org/c/gocb/+/153918
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>