GOCBC-1111: On status 200 parse any search errors from the status object
Motivation
-----------
Search can reply with a status code of 200 but still have errors which
are recorded in the status object. At present we try to parse these
errors from the top level object instead.

Changes
--------
Update search response json parsing so that the errors field is parsed
from within the status object.

Change-Id: I2e4caaf5ee955fba04b9679a866020c0e5872fa0
Reviewed-on: http://review.couchbase.org/c/gocb/+/155195
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>