GOCBC-1136: Cache metrics attributes map
Motivation
----------
Benchmarking is showing that gocb HEAD is allocating a much higher
number of bytes per op than 2.2.4. This is primarily due to the fact
that we allocate a new map every operation to send attributes (service
and operation name) to the meter. We should cache these maps instead,
the map should remain quite small due to being limited by the number
of operations we actually expose.

Changes
-------
Add a metrics wrapper object to hold the SDK meter, which also
an attributes sync map cache, used when creating attributes.
Update references to meter to now refer to the wrapper.
Don't use the cache or create any attributes when noop meter is
being used.

Change-Id: I5ce3bd96d7e09b7c8877de7df2ec39ccf7635d9e
Reviewed-on: http://review.couchbase.org/c/gocb/+/157026
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>