GOCBC-1140: Only set views group field if option is set to true
Motivation
----------
When a user sets Reduce to true on ViewOptions then we will always
set a value for group in the query payload. This means that if a
user uses GroupLevel then the query will fail as group and group_level
cannot not be set.

Changes
-------
Only set group in the view query payload if Group is true on the
options block.

Change-Id: I9929a167ad464268690ff9d734f3698d04b86f6d
Reviewed-on: http://review.couchbase.org/c/gocb/+/158552
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>