GOCBC-1168: Rework asyncopmanager to return errors at send time
Motivation
----------
Cluster level HTTP operations can hang indefinitely if gocbcore
returns an error on the send side of the operation, e.g. an
invalid arguments error. This is because we always call
wait on the asyncopmanager.

Changes
-------
Update asyncopmanager new function to accept the context.
Update asyncopmanager wait function to work the same as kvopmanager,
that is it takes an op and error and returns immediately if there
is an error.

Change-Id: Id6406c671c1bb9426d05d534d259d8fc69157654
Reviewed-on: http://review.couchbase.org/c/gocb/+/160627
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>