GOCBC-1171: Send fts not bounded consistency as empty string
Motivation
-----------
When the user sets a scan consistency of non bounded for a search
query then we send "not_bounded". This is wrong and it should be
sent as an empty string.

Changes
-------
Update search options to set scan consistency to an empty string
in the JSON payload when not bounded is set.

Change-Id: Iadc17a84a41ed4aa0ecb3143587b1f7b21ff4d4e
Reviewed-on: http://review.couchbase.org/c/gocb/+/161345
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>