GOCBC-1183: Return empty pool if no certs and skip verify is false
Motivation
----------
When a user provides no certificates and does not set skip verify
to true then connections made with the couchbases should fail to
connect. When skip verify is set then we return nil from the
gocbcore tls provider callback, if skip verify is not set then
we return the certificates that the user has set. If the user
has not set certificates then this value is also nil. This means
that we allow secure connections to be made with no certificates
and skip verify set to false.

Changes
-------
If skip verify is false and the user has set no root CAs then
return an empty CertPool.

Change-Id: Iaf3983566644af10c1d0e90591a9c1fa3f6c0c3e
Reviewed-on: http://review.couchbase.org/c/gocb/+/163346
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>