GOCBC-1185: Update Queue and Set to look for doc exists error
Motivation
----------
In the queue and set datastructures we check for a cas mismatch
error case on the mutation step of the pop and remove operations.
This is wrong as, due to a bug which would require a backward
incompatible fix, mutate in returns an exists error.

Changes
-------
Update Queue and Set to look for doc exists error during the pop
and remove operations.

Change-Id: I46b185cb56b9dd02e1779ffb206637241fc1fc91
Reviewed-on: http://review.couchbase.org/c/gocb/+/164273
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>