GOCBC-1186: Don't exit logging meter logger loop if nothing to record
Motivation
----------
In the LoggingMeter we do not log anything if there's nothing to log.
However, when this happens we exit the recorder loop. We should just
reloop and leave the loop running. A side effect of this is that
sometimes when the logging meter is exiting then the "no work"
check will have ended the loop early leaving the exit channel unread
which then hangs.

Changes
-------
Reloop in the logging meter is there is nothing to record, instead
of exiting the loop.

Change-Id: If75273762d81ce2406f244e3780f55334db4a91f
Reviewed-on: http://review.couchbase.org/c/gocb/+/164356
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>