GOCBC-1187: Don't output anything for empty aggregate meter groups
Motivation
-----------
During the logger output routine the SDK resets the recorder groups
which means that bins for the histograms are 0'd but are still
logged for this service/operation pair when they have no data.

Changes
-------
Only log a recorder's data if the total count of ops in the period
is greater than 0.

Change-Id: I89c9e32f457d42a92079368432e4f971e0f80839
Reviewed-on: http://review.couchbase.org/c/gocb/+/164360
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>