GOCBC-1154: Add support for rate limit errors
Motivation
-----------
With respect to the changes introduced in the servere,
all SDK's need to add logic to handle the newly introduced
rate limits such that the client can identify that these are
temp failures and take appropriate actions.

Changes
-------
Add RateLimitFailure and QuotaLimitFailure errors.
Add functions to mgmt_http and cluster_searchindexes to detect
and raise limit/quota errors.

Change-Id: I9f1c66bea0a8d55e33bfe111515318ac1cb59058
Reviewed-on: http://review.couchbase.org/c/gocb/+/164433
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>