GOCBC-1190: Add functionality to send queries to specific endpoints
Motivation
----------
A requirement of query in transactions is that all queries in a
transaction must go to the same node. We need to add a way to
facilitate doing that.

Changes
-------
Add internal type for query result and add endpoint to it.
Add endpoint to the internal field of  options.
Wire into gocbcore result/options endpoint fields.

Change-Id: I82485119104fc9c80571212c08b1097715273605
Reviewed-on: https://review.couchbase.org/c/gocb/+/165076
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>