GOCBC-1200: Rename rate limit errors
Motivation
----------
The RFC has been updated to change the prose for the rate
limit errors. From limiting to limited to be more consistent
with other errors. This is a breaking change but it is unlikely
that these errors are being used yet and this is really a bug.

Changes
-------
Change rate limit errors from limiting to limited.

Change-Id: Ib9adb8af4d401330ff3b98b5c019d64568a15cdd
Reviewed-on: https://review.couchbase.org/c/gocb/+/166782
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>