GOCBC-1217: Update search index quota limit error text check
Motivation
----------
Search have now provided us with a list of strings that we should
match against for rate/quota limit errors. We are mostly ok but
need to update the string we check against for number of indexes
exceeded.

Changes
-------
Update the string that we match against for checking when the
number of search indexes has been exceeded.

Change-Id: I3bd18b6b6c73c6503f629e8a4a1b46cf24f95e31
Reviewed-on: https://review.couchbase.org/c/gocb/+/169476
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>