GOCBC-1227: Update query for GetAllIndexes
Motivation
----------
When no collection or scope names are set the query that we use
for GetAllIndexes can match on collection names from other buckets
where the collection name is the same as the given bucket name.
We need to further restrict the query to only match on the given
bucket name.

Changes
-------
Update query for GetAllIndexes when no collection or scope names
are provided.

Change-Id: I16e68eda05e882c6163ea1c4430e2c6938f85ae5
Reviewed-on: https://review.couchbase.org/c/gocb/+/169477
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>