GOCBC-1228: Update queries used for BuildDeferredIndexes
Motivation
----------
We currently have an edge case where we can end up not building
all of the indexes that the user expects. If a bucket has a primary
index and a collection of the same name also has a primary index
and both are deferred then we will end up building only one of
these. We should make it clear that if no collection name is
provided in the options then only indexes on the bucket itself
will be built and adjust queries accordingly.

Changes
-------
No longer use GetAllIndexes in BuildDeferredIndexes, instead using
slightly more restrictive queries depending on whether collection
name is set or not.

Change-Id: I3144fbe70743d536c5e871c3a3cd2d1b43f4c77c
Reviewed-on: https://review.couchbase.org/c/gocb/+/169583
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>