TXNG-141: Add transactions option to cleanup extra collections
Motivation
----------
We have changed the automatic cleanup behaviour so that we no
longer run cleanup against all buckets. Instead, we run cleanup
on any ATR locations that we encounter during transactions. To
augment this we should also allow users to declare extra collections
that should be cleaned up.

Changes
-------
Add CleanupCollections to TransactionsConfig.
Update transactions atrLocationsProvider to return the set of
locations that the user provides us with.

Change-Id: Ib7f1f222b40c535f15cbb5ab7f77dc83347360bc
Reviewed-on: https://review.couchbase.org/c/gocb/+/169615
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>