GOCBC-1239: Add support for DurabilityLevelUnknown
Motivation
----------
In some situations we need to be able to determine the difference
between a durability being not set and being none. At present
we cannot do this, we need to add a new durability level which
takes the zero value and default its value to none in all current
places that it is used.

Changes
-------
Add support for DurabilityLevelUnknown as the DurabilityLevel zero
value.
Update kvopmanager to default the durability level to none when
unknown is set and to perform translations of the durability
level to the memd durability level.

Change-Id: I6bc81e01a194638848a041b30ab8c41d8211ae1b
Reviewed-on: https://review.couchbase.org/c/gocb/+/170342
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>