TXNG-158: Add support for per transaction metadata collection
Motivation
----------
We don't support having multiple transactions objects per cluster,
this means that if a user wants to use a different metadata collection
to what is already set then we need a way to do it on a per
transaction basis.

Changes
-------
Add MetadataCollection to TransactionsOptions and wire it into
gocbcore.

Change-Id: Ib99a380c23b6efea4d4596a3c6db327446796bb2
Reviewed-on: https://review.couchbase.org/c/gocb/+/170452
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>