TXNG-167: Don't return error from cluster.Transactions()
Motivation
----------
It's a better UX and more in keeping with the rest of the SDK
API if Transactions() does not return an error. We should move
transactions initiation to Connect so that any errors are returned
there.

Changes
-------
Changes signature of Transactions to not return an error.
Move initiation of transactions object to cluster Connect.
Change-Id: Ie2ca75225fcaca20ccd8f8be904f00e13e116090
Reviewed-on: https://review.couchbase.org/c/gocb/+/170912
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>