GOCBC-1250: Add support for single query transactions
Motivation
----------
As an earlier part of the transactions integration effort we
removed single query transactions. We should add them, integrated
into existing cluster/scope level queries.

Changes
-------
Add AsTransaction field to ClusterOptions.
Add singleQuery function to Transactions object, which is called
when TransactionsOptions set in the options passed to Query on
Scope or Cluster.
A couple of small tweaks to queryWrapper to match spec.
Add transaction id field to QueryResult to signal that calls to
Err and Close need to return transaction errors.

Change-Id: I4a42ba0537708e8c1fc76cc2923534e743970fa3
Reviewed-on: https://review.couchbase.org/c/gocb/+/171171
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>