GOCBC-1267: Include default collection in GetAllIndexes
Motivation
----------
When an index is created on the default collection in the default
scope then GetAllIndexes will not include that index if either scope
or collection name is not provided (or _default is used).

Changes
--------
Update that query used in GetAllIndexes so that default collection
is included when _default, or no, scope is specified.

Change-Id: Ia83b2c22b58de6e7484c410233f1274368009bb3
Reviewed-on: https://review.couchbase.org/c/gocb/+/173665
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>