GOCBC-1272: Don't drop connected buckets in tests
Motivation
----------
In a few tests we create a new bucket and connect to it, at the end
of the test we drop the bucket. This means that test logs end up
with noise in them from the SDK attempting to reconnect to the
dropped buckets.

Changes
-------
Instead of crewating new buckets drop any existing indexes so that
we can use the global bucket.

Change-Id: I3c1a7054365afc352d8865393819c91f02c81c9c
Reviewed-on: https://review.couchbase.org/c/gocb/+/175385
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>