GOCBC-1297: Use finalErrorToRaise txn state bits
Motivation
----------
Gocbcore has now moved to using finalErrorToRaise to indicate
what final error should be raised from the core loop. This will bring
us closer to spec and simplify the core loop.

Changes
-------
Update core loop to get the final error to raise from
the attempt.
If error is document not found then allow the raw error to be
raised to the lambda from get.
Mask the underlying error cause within TOF so that users don't
incorrectly match on errors when they shouldn't.

Change-Id: I77c5ee2ab02f4c3089f28e944265f19fe6c42466
Reviewed-on: https://review.couchbase.org/c/gocb/+/175556
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>