GOCBC-1213: Add support for ChangePassword to user manager
Motivation
----------
The server supports an endpoint changing the password for the
authenticated user used to make the request. We should add support
for this endpoint to the user manager. We should add documentation
that usage of the function will invalidate the SDK instance due to
future requests being subject to auth errors.

Changes
-------
Add support for ChangePassword to user manager.

Change-Id: I7348aec36443d35a1999fec731b491b538e9f14f
Reviewed-on: https://review.couchbase.org/c/gocb/+/176155
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>