Update internal txns API to not expose gocbcore
Motivation
----------
Within the FIT performer it would be very useful if the transactions
performer did not have a dependency directly on gocbcore. This means
that we need to do a bit of wrapping and aliasing to remove gocbcore
from the APIs used.

Changes
-------
Type alias TransactionErrorReason.
Update lost and regular cleanup initialisation to require gocb types
which wrap the gocbcore types.

Change-Id: Ib8647efb5ac46237f638a875f7fe39e351ab73d8
Reviewed-on: https://review.couchbase.org/c/gocb/+/178288
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>