GOCBC-1341: Add support for config profiles
Motivation
----------
We want to introduce a feature that allows our SDKs to be quickly
configured for some specifics use-cases. The first profile we will
create will be for "developer/high latency" environments and will
change the default timeouts to work in "high latency" environments.

Changes
-------
Add an internal developmentProfile which is an instantiation of
ClusterOptions.
Add an ApplyProfile function to ClusterOptions.

Change-Id: Ibaac474246eb1a84958a7e48dd519944184c14c7
Reviewed-on: https://review.couchbase.org/c/gocb/+/179446
Reviewed-by: Jack Westwood <jack.westwood@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>