Use unique names for scopes in eventing mgr tests
Motivation
----------
In the eventing manager tests we use the same name for the scope
in every test. As a part of each test we create the scope and then
drop it in a defer. Due to eventual consistency the scope from one
test can still exist when the next test attemps to create it.

Changes
-------
Use unique names for scopes in eventing mgr tests.Use unique names for scopes in eventing mgr tests

Motivation
----------
In the eventing manager tests we use the same name for the scope
in every test. As a part of each test we create the scope and then
drop it in a defer. Due to eventual consistency the scope from one
test can still exist when the next test attemps to create it.

Changes
-------
Use unique names for scopes in eventing mgr tests.

Change-Id: I1f927ccc0d72535ac83d162af90d69b8d2a9f6ee
Reviewed-on: https://review.couchbase.org/c/gocb/+/184246
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>