GOCBC-1373: Add query context support to transactions
Motivation
-----------
In a server version we now need to send query context as a part of
any queries, so we need to add support to transactions.

Changes
-------
Update transactions involving queries to send a query context if
scope if set on the QueryOptions.
Cache the scope and use it for further operations in that transaction.
Detect if a query error is feature not available and if so
rewrap the error if error code 1197 is responsible.

Change-Id: I211b68b62f21dbe3833a3f1bda2483d726e3f626
Reviewed-on: https://review.couchbase.org/c/gocb/+/184528
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>