GOCBC-1404: Use connstring rather than Server in tests
Motivation
-----------
Some of our tests use the raw connection string passed in by the
user. This is an issue when a certs path is specified because we
don't pick up the "couchbases://" prepending that that codepath
adds. We should refer to connstring rather than Server.

Changes
-------
Update tests using the raw connection string to use the
connstring.

Change-Id: I17bf4d7e55918df2a71afde4c119d952bb0c1e24
Reviewed-on: https://review.couchbase.org/c/gocb/+/189604
Reviewed-by: <matthew.bray@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>