Don't check number of spans in tests
Motivation
----------
In the tests we check the number of cmd and dispatch spans for
kv ops. This is a problem when the server has updated the config
whilst an op is queued for dispatch as it causes extra spans to
be added. We should not check the number of spans in the tests,
only validate the spans that are there.

Changes
-------
Update kv op span checking to not verify the number of cmd and
dispatch spans.

Change-Id: I969f5469d202df9d6e5d28b5f795d289288b7ca4
Reviewed-on: https://review.couchbase.org/c/gocb/+/189605
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>