Remove support for range scan
Motivation
----------
The specification for the implementation of range scan has changed
since we implemented it as a part of 2.6.x release. The 2.7.0
release will contain the updated version of range scan. As it has
been marked as volatile in the 2.6 release line we should remove
it from the API.

Changes
-------
Remove range scan from the API.

Change-Id: Id4a63931434261e191bc5f2e47d209b905813806
Reviewed-on: https://review.couchbase.org/c/gocb/+/193088
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>