GOCBC-1419: Add zapcore.Core implementation for PS logging
Motivation
-----------
gocbcoreps uses zap logging, which means a *zap.Logger needs to be
provided for any logging. As zap.Logger is a struct this means that
we have to create a zap logger using a custom core which can forward
into our Logger interface.

Changes
-------
Add a zapcore.Core implementation.
Register it with a zap.Logger and pass into gocbcoreps.

Change-Id: Ifee6ecd6012f1dfd5818738f91d5e38bff5f8e78
Reviewed-on: https://review.couchbase.org/c/gocb/+/193221
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>