GOCBC-1444: Return error if mutation state invalid in range scan
Motivation
----------
When we parse the mutation state for a range it's possible for
it to contain two tokens for the same vbucket. If the tokens don't
also have the same vbucket uuid we should return an error.

Changes
-------
Return error if mutation state contains two tokens for the same
vbucket id but different uuids.

Change-Id: I63e1785bb0a8b3b5ba3c3b1e2dc05e1badc5ac59
Reviewed-on: https://review.couchbase.org/c/gocb/+/193630
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>