GOCBC-1492: Update how we assign errors in setting kv durability
Motivation
-----------
When we set durability level options on an op after setting the
value then we will swallow any transcoding errors as we assign
the error value straight into the op manager error, even if that
value is nil.

Changes
-------
Update op manager so that the error is only assigned if it is
non-nil after translating to memd.

Change-Id: I8c620935ce7950b04f5f5e071aef0f7ae555b526
Reviewed-on: https://review.couchbase.org/c/gocb/+/198211
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>