GOCBC-1514: Update PS BuildDeferredIndex return and state handling
Motivation
----------
BuildDeferredIndexes is not returning any of the index names from
the PS response. It is also returning a different value for the
State field compared to classic.

Changes
-------
Update index state to internally be typed and add translation to
the ps query index provider.
Update return from BuildDeferredIndexes to contain the index names
from the PS response.

Change-Id: I7c2737af768b238b6246f1b627255b8fbe6b87cf
Reviewed-on: https://review.couchbase.org/c/gocb/+/199333
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>