GOCBC-1518: Update timeout error handling in couchbase2
Motivation
-----------
When an operation times out in flight we are not returning the
error as a TimeoutError, but a GenericError.

Changes
-------
Update handleRetriableRequest to create timeout errors with the
gocb translated error as InnerError rather than the grpc error.
Update kv op manager to not call enhance error, as the error
has already been translated and enhanced.

Change-Id: I11a1ae614fc6f7641b984c6c8dbeba6ccb13143f
Reviewed-on: https://review.couchbase.org/c/gocb/+/199366
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>