GOCBC-1530: Update append/prepend to translate NotStored to NotFound
Motivation
----------
There are a few places where the server can return a not stored
status to us for kv ops. We already translate most of these to
not found errors, however append/prepend were missed. Changing
this in gocbcore would be a breaking change but we do not expose
ErrNotStored in gocb so we can still do the translation.

Changes
------
Update append/prepend to translate NotStored to NotFound.

Change-Id: I65a499a1edd111f9c7ebef9d4aa607670963bab6
Reviewed-on: https://review.couchbase.org/c/gocb/+/200987
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>