GOCBC-1580: Expose set of constants for geo distance sort units
Motvation
----------
We expose geo distance sort units as a string, which is not very
discoverable. We should add a set of constants that can be used with
this field. We cannot change the field itself from a string without
being a beraking change.

Changes
-------
Expose set of constants for geo distance sort units.

Change-Id: Ifc7213a539be50f3cbeb3cec4079dd963275bcb6
Reviewed-on: https://review.couchbase.org/c/gocb/+/203754
Tested-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>