GOCBC-1597: Convert 'non existent bucket' error message to ErrBucketNotFound
Motivation
==========
In newer server versions, attempting to do a bucket management operation with a non-existent bucket results in a 404 error with the message "Attempt to access non existent bucket". This is not being converted to ErrBucketNotFound by gocb.

Changes
=======
In bucket management operations, if the response status is 404 & the error message contains 'non existent bucket', return ErrBucketNotFound

Results
=======
All tests pass

Change-Id: I610a0d2951da4b8908e58b9621790d57cca09aca
Reviewed-on: https://review.couchbase.org/c/gocb/+/205310
Tested-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>